chore(models): Move review/delta-models to devtools-api models
These are generated via malli->json-schemas. The delta-model is complemented by sorting info and optional "refactorableFn"
