+++ src/code-health-monitor/details/view.ts
@@ -4,8 +4,9 @@ import { refactoringButton } from '../../codescene-tab/webview/refactoring-compo
 import { issueToDocsParams } from '../../documentation/commands';
 import Telemetry from '../../telemetry';
 import { commonResourceRoots, getUri, nonce } from '../../webview-utils';
-import { ChangeType, isDegradation } from '../model';
+import {  isDegradation } from '../presentation';
 import { DeltaFunctionInfo, sortIssues } from '../tree-model';
+import { ChangeType } from '../../devtools-api/delta-model';
 
 export function register(context: ExtensionContext) {
   const viewProvider = new CodeHealthDetailsView();
